+++ util-core/src/test/scala/com/twitter/util/TwitterDateFormatTest.scala
@@ -20,22 +20,4 @@ class TwitterDateFormatTest extends AnyFunSuite {
   test("TwitterDateFormat should allow Y when quoted") {
     TwitterDateFormat("yyyy 'Year'")
   }
-
-  test("TwitterDateFormat should stripSingleQuoted") {
-    import TwitterDateFormat._
-    assert(stripSingleQuoted("") == "")
-    assert(stripSingleQuoted("YYYY") == "YYYY")
-    assert(stripSingleQuoted("''") == "")
-    assert(stripSingleQuoted("'abc'") == "")
-    assert(stripSingleQuoted("x'abc'") == "x")
-    assert(stripSingleQuoted("'abc'x") == "x")
-    assert(stripSingleQuoted("'abc'def'ghi'") == "def")
-
-    intercept[IllegalArgumentException] {
-      stripSingleQuoted("'abc")
-    }
-    intercept[IllegalArgumentException] {
-      stripSingleQuoted("'abc'def'ghi")
-    }
-  }
 }